Parts suppliers racks validations

+++ Backend/handler/parts.py
@@ -1,5 +1,6 @@
 from flask import jsonify
 from Backend.DAOs.parts import PartDAO
+from Backend.DAOs.supplies import SuppliesDao
 
 
 class PartHandler:
@@ -45,16 +46,32 @@ class PartHandler:
         if result:
             return jsonify(self.mapToDict(result))
         else:
-            return jsonify("Not Found"), 404
+            return jsonify(f"Did not find a part with id: {pid}"), 404
 
     def insertPart(self, data):
         # print(data)
+        if len(data) != 4:
+            return jsonify("Did not receive the correct amount of information needed for a Part record."), 400
 
         # looking up the values in the dict
-        name = data['Name']
-        color = data['Color']
-        material = data['Material']
-        msrp = data['msrp']
+        try:
+            name = data['Name']
+            color = data['Color']
+            material = data['Material']
+            msrp = data['msrp']
+        except:
+            return jsonify("Error: Invalid argument names!"), 400
+
+        if not isinstance(name, str):
+            return jsonify(f"Error: The inputted name '{name}' is not a string!"), 400
+        if not isinstance(color, str):
+            return jsonify(f"Error: The inputted color '{color}' is not a string!"), 400
+        if not isinstance(material, str):
+            return jsonify(f"Error: The inputted material '{material}' is not a string!"), 400
+        if not isinstance(msrp, float) and not isinstance(msrp, int):
+            return jsonify(f"Error: The inputted msrp '{msrp}' is not a valid number!"), 400
+        if msrp <= 0:
+            return jsonify(f"Error: The market sale retail price (msrp) entered has to be bigger than 0"), 400
 
         if name and color and material and msrp:
             dao = PartDAO()
@@ -65,19 +82,52 @@ class PartHandler:
             return jsonify("Unexpected attribute values."), 400
 
     def deleteByID(self, pid):
+        # Can't delete a part that is being supplied
         dao = PartDAO()
+
+        in_stock = dao.inStock(pid)
+        if in_stock:
+            return jsonify("Cannot delete this part, since it is currently being supplied."), 400
+
+        # Can't delete a part that is being referenced in a transaction
+        in_transaction = dao.inTransaction(pid)
+        if in_transaction:
+            return jsonify("Cannot delete this part, since it is being referenced in a past transaction."), 400
+
+        being_stored = dao.beingStored(pid)
+        if being_stored:
+            return jsonify("Cannot delete this part, since it is currently being stored in a warehouse."), 400
+
         result = dao.deleteByID(pid)  # if this is null, then the part with that id doesn't exist
         if result:
             return jsonify(f"Deleted part with id: {pid}"), 200
         else:
-            return jsonify("Not Found"), 404
+            return jsonify("Did not find a part with id:", pid), 404
 
     def updateByID(self, pid, data):
+
+        if len(data) != 5 and len(data) != 4:
+            return jsonify("Did not receive the correct amount of information needed for a Part record."), 400
         # looking up the values in the dict
-        name = data['Name']
-        color = data['Color']
-        material = data['Material']
-        msrp = data['msrp']
+        try:
+            name = data['Name']
+            color = data['Color']
+            material = data['Material']
+            msrp = data['msrp']
+        except:
+            return jsonify("Error: Invalid argument names!"), 400
+
+
+        if not isinstance(name, str):
+            return jsonify(f"Error: The inputted name '{name}' is not a string!"), 400
+        if not isinstance(color, str):
+            return jsonify(f"Error: The inputted color '{color}' is not a string!"), 400
+        if not isinstance(material, str):
+            return jsonify(f"Error: The inputted material '{material}' is not a string!"), 400
+        if not isinstance(msrp, float) and not isinstance(msrp, int):
+            return jsonify(f"Error: The inputted msrp '{msrp}' is not a valid number!"), 400
+        if msrp <= 0:
+            return jsonify(f"Error: The market sale retail price (msrp) entered has to be bigger than 0"), 400
 
         if pid and name and color and material and msrp:
             dao = PartDAO()
@@ -85,6 +135,6 @@ class PartHandler:
             if flag:
                 return jsonify(data), 200
             else:
-                return jsonify("Not Found"), 404
+                return jsonify(f"Did not find a part with id: {pid}"), 404
         else:
             return jsonify("Unexpected attribute values."), 400
